Cambios en menu

+++ src/main.cpp
@@ -1,7 +1,7 @@
 #include "../Menus/Menu.h"
 
 int main() {
-	Menu< char, int > *graphMenu = new Menu< char, int >();
+	Menu< char, double > *graphMenu = new Menu< char, double >();
 	graphMenu -> principalManu();
 	return 0;
 }
